Move feature_calculation to nodes module
