--- src/machine_learning/data_processors/__init__.py
@@ -1,3 +1,0 @@
-"""Module for data transformers"""
-
-from .feature_calculation import feature_calculator
